--- a/docs/configuration/logs.md
+++ b/docs/configuration/logs.md
@@ -128,7 +128,14 @@ To write the logs in async, specify `bufferingSize` as the format (must be >0):
 ```toml
 [accessLog]
 filePath = "/path/to/access.log"
-bufferingSize = 262140
+# Buffering Size
+#
+# Optional
+# Default: 0
+#
+# Number of access log lines to process in a buffered way.
+#
+bufferingSize = 100
 ```
 
 To filter logs you can specify a set of filters which are logically "OR-connected". Thus, specifying multiple filters will keep more access logs than specifying only one: