@ldez 2018-05-16
@ryarnyah thanks for  your work.

If you could not rebase and address the review comments in terms of additional commits, this is very useful for us because it facilitates the review.

https://github.com/containous/traefik/blob/master/CONTRIBUTING.md#content
> - Address review comments in terms of additional commits.
> - Do not amend/squash existing ones unless the PR is trivial.