@ryarnyah 2018-05-16
@ldez That's what i did. I was just asking if i must squash all of them after reviews :D